System Re-Factor Push 1
Added the new interface system VaporArcade. I also exposed the
ApplicationSettings class through the VaporArcade class as a model for
the Business Manager to Business Object relationship. An application
path InterfaceDirectory was added to store the location for the
different User Interfaces. Began complete overhaul of chat system.
Working on MainLobby an UDPSocket based derived class of iLobby. This
class will send UDP Datagram broadcast for the main lobby chat system.

NOTE: If building on windows you will need to direct the compiler to the
Apple Bonjour development libraries on your machine. These need to be
GCC compatible (.a not .lib). If compiling on Linux everything should
still work with the libavahi interface.